Updated the remove button to work, modules now stay on reload
Updated structure so that at Planner.js level there is one list of all modules. Each module will now be tagged with the year of the users choice and each year will filter for it. used uuid to give each module a unique id to delete it when necessary. Need to add a remove all button. Now modules will also be stored in localStorage.